Add logic to monitor VPN endpoints
We need a way to be able to monitor the status of a VPN (and be able to
set hosts dependent on that check).

This commit adds a new `check_ping_external` command because we do not
want to create a separate Icinga host for each VPN gateway, but the
default `check_ping` command will monitor what is set for host_name,
which is dependent on creating an Icinga host.

This can also be used to monitor any external machine if required.