Monitoring should default to true
We should default things like monitoring to true and disable them in
environments where we don't want them. This means that our
production configuration file is as small as possible.

I've switched the monitoring off in the specs so the behaviour is
actually consistent with what we already have, but really we should
be testing this parameter.