Serve robots.txt from router nginx.
Serve robots.txt from router nginx instead of the static app.
So that we can change the contents depending on environment.

The location block is inside the location block for / so that if the
vhost is password protected the robots.txt file will be as well.

This takes the robots.txt file from the static application:
https://github.com/alphagov/static/commit/577ac17308cd6b0ef004c9051f75e7be5902a958