Fix Puppet duration stats in Graphite
We used to send stats on the amount of time Puppet is taking to run
to Graphite.

This broke when we changed machine hostnames because we're not using
the correct facter fact in `govuk_puppet`. The machine hostnames are
too long so we created `fqdn_metrics` fact to handle the long hostnames.

This needs to be a template rather than a file because we're using
ERB to insert the value of a facter fact.