Add defined type to ping 'external' hosts
Add a defined type that creates an Icinga check to ping hosts for which
we have not defined an `Icinga::Host` resource.

This is useful to ping hosts other than our VMs. Our first use case will
be to use this to monitor VPNs by pinging the network gateway on the
other side of the VPN tunnel.

We've added a new Icinga command, `check_ping_external`, because the
existing `check_ping` command always uses the `host_name` parameter
passed into `Icinga::Check`, which must be a machine that Icinga knows
about (defined as a `Icinga::Host` resource), otherwise Icinga will
error. Creating a new command allows us to pass the host to ping as an
argument to the command.

The thresholds for the average round trip threshold are low because we
expect to use this check for monitoring geographically local hosts.