Monitor VPNs by pinging network gateways
Monitor our VPN tunnels by pinging the network gateway on the other side
of each tunnel.

We've feature-flagged this to disable it in Staging and Production until
we've tested it in Integration to avoid false alarms in Production.

Note that although EFG doesn't exist in Integration, we still
monitor the VPN tunnel as:

a) it means we don't have to override the Hiera data here for
   Integration

b) it keeps our config as close as possible to Production

c) the VPN and network already exists for EFG in Integration, which
   allows us to test our VPN configuration in that environment