Manage Enterprise Finance Guarantee certificate
This certificate has never been managed by Puppet, which is a bad
thing. It means that if we lost the machine we would not be able
to provision it simply.

The certificate in the YAML files in this commit are a self-signed
certificate for www.sflg.gov.uk which I've just generated locally.