Exported resource for DR host dependency
Create an exported resource which only applies to DR hosts using the
dr_node fact. This is so that if the appropriate VPN gateway DR service
check goes down we will not receive notifications or alerts for any DR
hosts behind this gateway. We assume that if the gateway is not
responding then the VPN is down, and thus monitoring is unable to reach
any of the hosts.