Add snakeoil X.509 RSA TLS certificate/key pair
...to be used during spec tests.

In preparation for using the `validate_x509_rsa_key_pair()` function
from the Puppet stdlib, which will error if the certificates we use
during testing are not valid. A self-signed certificate using
`localhost` as the common name is sufficient to pass the validation.

I'm not keen on the relative path used in
`spec/fixtures/hiera/real_data.yaml`, but the only alternatives were to
override the Hieradata key in
`modules/router/spec/classes/router__assets_origin_spec.rb`, which would
add a lot of [complexity][] by requiring the addition of an 'rspec'
Hiera backend, or add a snakeoil cert/key pair in our real Hiera
`common.yaml`, which I absolutely do not want to do lest we accidentally
use it on a Production system.

[complexity]: https://github.com/bobtfish/hiera-puppet-helper#advanced