Bugfix display_name in Icinga
Looking up this Facter fact inside the template means that the fact
will always evaluate the value it is on the monitoring machine, which
changes the display name of every machine to be `monitoring-1`.

Looking up the fact in the defined type means that it's evaluated
on the client host.