Add the option to define parents for hosts
This would allow us to add the parent definition to a given host, so
that if the parent is down (for example a VPN gateway) then the
notifications and/or checks for a host will be ignored (in theory).

The `hostname` and `fqdn` facts were not required in these tests, and add tests
for the if statement within the icinga::host define.