Add UFW rule for asset slave using NFS
The asset slave machines need to connect to the asset master using
NFS.

asset-slave-1 is currently allowed to connect because of the broad
rule in `s_asset_base` which allows all backend machines to connect
to the asset-master.

I'd rather not extend that broad rule to include asset-slave-2 (running
on a different IP range in the disaster recovery organisation) so instead
I've set up new UFW rules which are different:

- They only apply to the asset_master rather than the asset_base class
- They're explicit about being for the asset_slave only

In an ideal world I'd like to limit these rules by port number too, but
I don't know enough about what ports NFS uses (it seems to use lots of
high numbered ports).