Create network_checks defined type
This creates a defined type which creates a host and a ping check on top
of that host. This can be used to create simple monitoring for any
networking devices. The immediate requirement was to be able to monitor
vDC gateways, and create them as hosts so they could be used as parents
for any "children" nodes behind them.