Add fields emitted by CDN logs monitor
The `cdn_backend` indicates which CDN backend handled the request, and
is one of a set of possible string values, so we don't want it analyzed
(it's similar to the `@tags` field).

The `last_success` field indicates the time at which we last recorded a
successful access to the resource.  This is populated only for records
which are reporting a failure of an access to a resource which has
previously been known to succeed.

It's a date, so it's good to tell elasticsearch this rather than rely on
it guessing correctly.