Run full attachment sync later in staging
Currently this runs at both 4am in staging and production. However, the
sync of attachments from production to staging has finished at these
times recently:

- 06:13
- 05:54
- 05:51
- 05:55

This means that the data sync will run after the full attachment sync,
which is bad because the asset master and asset slaves will be out of
sync in staging for most of the day.

Moving the full rsync 3 hours later means this doesn't happen.