Tweak attachment rsync cron
This commit changes this cron entry to be similar to the way we handle
draft virus scanning - half an hour later than the clean equivalent,
and with a different lockfile. I can't foresee any issues with running
the clean/draft jobs with different lockfiles (they should be able
to run at the same time).