Update postgresql module
Version 4.1.0 of the PostgreSQL module has support for the future parser,
which is something it would be nice if we could use.

https://github.com/puppetlabs/puppetlabs-postgresql/compare/3.4.1...4.1.0

The backwards incompatible changes in the 4.x series are because there's
no longer the ability to uninstall PostgreSQL using this module and it
doesn't manage firewall rules anymore. Neither of these are features we
need:

- We'd never uninstall PostgreSQL from a machine, we'd destroy the machine
  and provision a new one
- We don't rely on the PostgreSQL module to manage our firewall rules

The other change is that Puppet will now manage `pg_ident.conf` by
default. I don't know why we wouldn't want this file managed.