Replace `govuk_rbenv::all` with new parameter
The parameter `govuk_rbenv::rubies: all` does the same thing as
including `govuk_rbenv::all`.

I've moved this to YAML because it's machine configuration. I appreciate
that this means we're no longer configuring the rubies in
`govuk::node::s_ruby_app_server`, but I think that class is broken
and should be deprecated and eventually removed (for example, why
on earth are we including `mysql::client` on the draft cache
machines?

Functionally this should have no change.