Tidy .gitignore
The `.gitignore` file contains a number of files or directories that
are more appropriate in a user's global gitignore rather than in this
project-specific `.gitignore`.

For example, we are ignoring files generated by:

- bundler, depending on your configuration
- emacs
- geppeto (a Puppet IDE)
- rubygems
- IntelliJ
- `.DS_Store` files created by OS X

...and so on. Rather than clutter `.gitignore` with files specific to
each editor, leave it to developers to configure their global
`.gitignore`. This makes it easier for these files to be accidentally
committed, but then they could just as easily accidentally be committed
other repositories worked on by those users and maintaining `.gitignore`
files across all our repositories in this way is undesirable.