Increase Postgres backup partitions
In both Production and Staging the partitions for Postgres backups are
right on the edge of filling up in space, and presumably will continue
to rise as time goes by.

This change seeks to double the size of the disks so we don't have to
worry about this for a while.