Specify memory alert thresholds in MB
When tweaking these thresholds I found it tedious and error-prone to
edit values in bytes.

Instead, allow the memory alerting thresholds for applications to be
specified in MB.

I've chosen to use the SI definition of 'megabyte', i.e. 1000000 bytes,
because it makes it easier for us to review these thresholds on a
regular basis when comparing against our Graphite metrics, which store
the data in bytes.

I've also added some validation to ensure nobody accidentally sets the
value in bytes and some corresponding tests.

We've lost some precision when specifying existing thresholds, e.g. for
Whitehall, but given that we're working with megabytes and Whitehall
uses a magnitude greater in memory, that should not be an issue.

I wanted to test that the `icinga::check::graphite` resource was being
exported with the appropriate parameter, however it's not currently
possible to assert this using rspec-puppet:
https://github.com/rodjek/rspec-puppet/issues/113

I've also taken advantage of class parameter defaults to remove the
conditional block that sets the `nagios_memory_*_real` variables.