Allow rotate user in logrotate config
We have come across an issue where logfiles are not being rotated
because the parent directory is not root, and a different user is not
explictly defined.

This is the related error:

```
considering log
/data/vhost/signon.publishing.service.gov.uk/shared/log/production.json.log
error: skipping
"/data/vhost/signon.publishing.service.gov.uk/shared/log/production.json.log"
because parent directory has insecure permissions (It's world writable
or writable by group which is not "root") Set "su" directive in config
file to tell logrotate which user/group should be used for rotation.
```

This change seeks to add the option within the defined type, and only if
both user and group options are specified add in the required directive.