Allow nginx logs to rotate even if empty
There have been some instances that logs haven't properly rotated for
nginx, causing the logs to grow in size very rapidly.

This allows the logs to rotate even if there was an issue and any logs
below the main log are empty, which should hopefully resolve the problem
of the main log growing, which can become impacting to the running of
the system.