Remove email-alert-monitor
email-alert-monitor was planned to be an application to monitor the
sending of email-alerts. Currently there are no more plans for this to
be a stand-alone application, so it can be removed from the system. The
application has been deployed to integration and staging, but doesn't
do anything (the repo is a blank Rails app)[1].

Separately, the repository will be retired to gds-attic and the app
will be removed from the `development` repo on GDS internal GitHub.

Note that we are currently working on a system to monitor certain email
alerts. This will be a script that tells Icinga to alert when a
publication hasn't triggered an email alert[2].

As discussed with @rboulton.

[1] https://github.com/alphagov/email-alert-monitor
[2] https://github.com/alphagov/email-alert-monitoring