Retire email-alert-monitor
email-alert-monitor was planned to be an application to monitor the
sending of email-alerts. Currently there are no more plans for this to
be a stand-alone application, so it can be removed from the system. The
application has been deployed to integration and staging, but doesn't
do anything (the repo is a blank Rails app).

This application will be removed in
https://github.com/alphagov/govuk-puppet/pull/4053.