Set up email alert monitoring in Icinga
We want to verify that emails are sent for all drug alerts and medical
device alerts (https://www.gov.uk/drug-device-alerts).

This commit sets up a Jenkins job, which runs every hour. It also sets
up a Icinga "passive alert", which will notify us if the job has not
run successfully.

The jenkins job runs a script that actually checks if the emails have
been sent:

https://github.com/alphagov/email-alert-monitoring/pull/1