Add CONTRIBUTING.md
Richard Boulton pointed out that the Infrastructure team has not been
good at communicating that we've moved to a self-merge approach for
Puppet pull requests.

Dean Wilson suggested that we do it because Puppet is a repo which
has lots of people touching it, and it's nice to be able to control
when your changes are merged (and therefore deployed).

This file seems like the best way to document that.