Add docs directory
This directory will contain documentation.

The first bit of documentation is the bit of the opsmanual which
explains how to add a new app (opsmanual@9f3ad8e7870c757cfa8d8386e11a6f6981771dcc).
I've converted it from reStructuredText to Markdown.

Moving this documentation closer to the Puppet repo means it's more
likely to be kept up-to-date.