Backup processed CDN logs
We're going to start deleting historical raw CDN logs, which means
the processed files will be the only data we retain. We should
back them up.

I've forced the `processed_data_dir` parameter to be in hieradata
so that I can do a hiera lookup in `backup.yaml` and avoid some
duplication.

https://trello.com/c/xbsYTau0/33-back-up-processed-cdn-logs