Make integration work one hour earlier
The content team are using integration for training and they need
to have it working in the morning.

Specialist publisher currently has an issue where the database is in
an inconsistent state in integration between the times when:

- The database is upgraded
- The data sync complete job is run

Stephanie Richardson-Dreyer has agreed to having integration switched
on a little bit earlier.