Add Mapit to api load balancer
This is to enable us to use the new version of Mapit (initially for testing
 with dependant apps in Integration).

This part sets up the configuration so that both versions are available.

When ready, we'll change the config so that the Mapit URL points at the new
version. We'll then be able to test before switching it back.

https://trello.com/c/LrsfgWXN/30-test-the-new-version-of-mapit-in-integration