Define icinga checks for every consumer
Currently we define these checks in the app modules for a few consumers,
meaning they get run from the backend, and make requests against the
first rabbitmq node.

We get duplicate checks when services are clustered, and we don't have
checks defined for all consumers.

This changes the check to run on the rabbitmq nodes, against localhost,
and we define it in the same place we create service users, so the checks
are extended to every consumer.

There is still some duplicated checks because clustered nodes will return
the same result, but its easier to keep the configuration the same for
all of them.