Add Mapit to api load balancer
This is to enable us to use the new version of Mapit (initially for testing
with dependant apps in Integration).

This part sets up the configuration so that both versions are available (ie
via the backend or api loadbalancers).  The new Mapit server type is not
available in staging or production yet.

https://trello.com/c/LrsfgWXN/30-test-the-new-version-of-mapit-in-integration