Make upstream apps aware of secure status
Sends the `X-Forwarded-Proto` header to nginx upstream so that apps can
make decisions based on whether a page is being served over http or
https. Relies on the header being passed through on the per-box nginx
to the unicorn processes. It’s not currently blacklisted with
`proxy_hide_header`.

Will allow us to set `config.force_ssl` in Signon.