Make port use an integer value, not implicit string conversion
This is an early step to support future/parser puppet 4. Our rspec
tests currently specify (correctly) that port is an int. Our manifest
should reflect that choice of type.

We also shouldn't be running our apps on priviledged ports so we set
the lowest allowed port to be 1025.

This will require apps using strings to be converted over to ints
when we start to run puppet 4. Those will be done in a seperate PR.