Add a vdc fact to our test runs to make them pass
When testing under the future parser we see issues in
icinga/manifests/client.pp under
"if ($::vdc == 'licensify') or ($::vdc == 'efg') or ($::vdc =~ /^*_dr$/) {"
if the vdc is undefined.

Adding this fact doesn't fix the tests but it does move the problem along from
'Left match operand must result in a String value. Got an Undef Value.'
to Could not find template '' which is an issue from an upstream module we'll
deal with in isolation.