Add parameter to error on plaintext HTTP requests
Generally we prefer to redirect from HTTP to HTTPS to provide a good user experience,
but for APIs and machine-only apps we can serve errors for HTTP connections.

The logic I've implemented is as follows:

- error_on_http true: https_redirect is ignored, errors are served
- error_on_http false: https_redirect decides whether to redirect or serve the app

According to the spec [1]:

> A server MAY indicate that a client request can not be completed
> without TLS using the "426 Upgrade Required" status code, which MUST
> include an an Upgrade header field specifying the token of the
> required TLS version.

I am unable to add a header to an error response using `add_header`
so we're going to have to violate the RFC for now (this does not
bother me).

Thanks to Jenny Duckett for finding the correct status code to use.

[1]: https://tools.ietf.org/html/rfc2817