Ensure GPG key is re-imported when changed
Currently, if we change the GPG private key in Hiera data, the new key
is not imported.

This change ensures that the old key is deleted to allow the new key to
be imported.

Since `refreshonly` is set to `true`, this command should only run when
the private key file is changed and so I have removed the `unless`
option as it's unnecessary.

The GPG fingerprint must now be a full fingerprint for the `gpg --batch
--delete-secret-key` to work; it errors if I try to supply a full
fingerprint.

When deploying this, I'll need to:

- [ ] Raise a corresponding PR in the deployment repo to change the
  GPG fingerprint to be full-length

- [ ] Remove the old secret key files using Fabric
  i.e. `/root/.gnupg/<SHORTFINGERPRINT>_secret_key.asc`