Allow some Elasticsearch clusters reboot in Prod
Allow the backend and API Elasticsearch clusters to reboot in Production
unattended overnight. They will only reboot if the cluster is healthy.

I haven't enabled this for the logs Elasticsearch cluster as the shard
reallocation, which is not disabled by the unattended reboots, takes
hours and I'd rather test it further in Staging first.