Add wal_e class
This commit looks to add and create WAL-E backups for Postgres to an S3
bucket.

It does two things:

 1. Continous archiving
 2. Daily base backup

The continous archiving is so easy to implement that I thought it would
be more work to create something that separately created WAL archives.
The daily base backup is to ensure a consistent base, and is managed
within /etc/cron.d to ensure flexibilty while we still have the other
backup system in place.