Move PARALLEL_JOBS env var to correct place
The `PARALLEL_JOBS` environment variable in the data sync Jenkins job
was in the wrong place.

Fixing this will allow our Integration sync job to complete before
office hours by running the sync jobs in parallel; currently it has been
overrunning until after 9:30am. It has been taking 1.5 hours longer
since the `mongo-pp` job was added in alphagov/govuk-puppet#3665.

According to the contextual help in Jenkins, `script-content` is used as
follows (note the last line):

    Execute a script file aimed at setting an environment such as creating folders, copying files, and so on.
    Give the script file content.
    You can use the above properties variables.
    However, adding or overriding environment variables in the script doesn't have any impacts in the build job.

So that this variable is taken into account by the build job, move it to
`properties-content`, which is the correct place to set environment
variables for the build.