Don't create app.pid for unicornherder procfile apps
As part of deploying a new version of Mapit, we want to use Unicorn Herder to
help us with zero-downtime deploys.

We were experiencing problems with this, as the pid file was being created by
the Unicorn Herder process (running as the super user), which meant that the
gunicorn processes couldn't access it (as they run under the app user).

We know that other applications use the 'procfile' app type which has
'--make-pid' defined, however we want to skip this in this instance so that
the first gunicorn process creates it instead.

https://trello.com/c/d82WTfeq/314-get-mapit-working-with-unicorn-herder-3-time-boxed-1-day