Allow jumpboxes to reboot unattended overnight
The reason we had not previously allowed the jumpboxes to reboot
overnight was to avoid exacerbating an incident if someone was
investigating an issue out of hours (the time when the unattended
reboots occur) and was logged off mid-incident because the jumpbox was
rebooted.

This script checks if there are any users logged on so that the jumpbox
can reboot overnight.

Note that there's still the potential that the jumpbox never returns
from the reboot, or that it's still rebooting when someone needs to log
in, however I think that that scenario is rare enough that it's not
something we should worry about. We also have a second jumpbox.