Don't analyse `fields.govuk_original_url` in logs
https://github.com/alphagov/govuk-puppet/pull/4135 was meant to mark
this field as not analysed, but used an incorrect field name. The
correct one is `govuk_original_url`.

From the original PR:

> The purpose of recording this is to be able to find the list of pages
which a particular app is involved in rendering, or the list of apps
involved in rendering a particular page. If the field is analysed we
can't do this, because a facet query will instead return the list of
components of urls.

Trello: https://trello.com/c/JVZPH4S9