Server uptime: Filter out null/zero values
The panel showing servers rebooted in the last 24 hours looks for
uptimes of less than 24 hours, meaning that it will show hosts that no
longer exist.

Fix that by filtering out zero or null-value metrics.

The issue is less likely to be an issue for the top 10 machines with the
highest uptime (longest running), but apply the fix there also for
completeness.