Use dedicated disk for Duplicity cache
Move the Duplicity cache onto its own disk, which will allow us to
monitor the disk space used by assets synced from the asset master
without the size of the cache directory interfering.

We perform offsite backups using Duplicity on the asset-slave-1
machines. Duplicity maintains a cache which is uses to know what is
already backed up when performing incremental backups. Otherwise, it
would have to retrieve the backups from the remote and decrypt them.

We recently started rsyncing /mnt/uploads from the asset-master-1 to
the asset-slaves and added an Icinga check (ee785db7) to ensure that the
free space on each machine corresponds (to confirm that the sync is
working correctly). Since the Duplicity cache directory exists on the
asset-slave but not the asset-master, this Icinga check is raising false
alarms.