keepalived: allow default auth password to be set
Add an `auth_pass` class parameter to `govuk_keepalived` that allows us
to specify an authentication password used by the VRRP protocol.

This password will be shared by all VRRP instances configured by the
`govuk_keepalived` module. I'm going to configure VRRP instances per
node type. So for example, if I configure two VRRP instances for the
`draft_cache` node type, both of those VRRP instances will have share
the same password. VRRP instances for other node types will use their
own passwords.

The reason for doing this is we're using `create_resources` to configure
the VRRP instances based on a hash and without this class parameter,
we'd need to specify `auth_pass` in the `deployment` repository where
our credentials are stored along with the rest of the VRRP instance
configuration since we're not performing deep merging of hashes in Hiera
data. I'd rather not have to store the full hash in the deployment repo,
so adding this class parameter is a compromise.

In reality, I think sharing the authentication password between VRRP
instances on the same node type is not a big risk  and for now we only
plan to have one VRRP instance per node type.

Also warn if instances are specified by the password is unset.