Streaming MySQL backups to S3
This commit strives to create incremental backups to an S3 bucket which
are no more than 15 minutes out of date.

Functionally it uses Percona Xtrabackup
(https://www.percona.com/doc/percona-xtrabackup/2.4/index.html),
using the innobackupex wrapper
(https://www.percona.com/doc/percona-xtrabackup/2.2/innobackupex/innobackupex_script.html)
and gof3r (https://github.com/rlmcpherson/s3gof3r) to take the stream
output and push to an S3 bucket.

It's recommend use is to take a base backup and then incremental backups
on top of that. It also compresses and encrypts on the fly.

Within the bucket, it creates a new folder for each set of backups, and
does this by the base backup echoing a date into a file, which the
incremental backup then references. This is so it splits up the daily
backups and hopefully makes it easier to restore.