Upgrade puppetdb to 2.0.0
This commit upgrades puppetdb on precise machines to 2.0.0.

2.0.0 requires a newer version of the JDK than 6. 6 seems to be
provided by the old puppetdb package so I've uninstalled it and manually
installed 7.

I've also modified the upstart file to allow puppetdb to start - the
change I've made here is similar to the change the package made to
its own init file (which we purge).