Use new MapIt in Staging
This is to switch Staging to using the new MapIt servers.  Once we have run
the set of tests detailed in the Trello card, we will have greater confidence
that the switch over in Production will go smoothly.  We envisage that the
environments will be different for a few days.

This supersedes this PR which we will redo:
https://github.com/alphagov/govuk-puppet/pull/4220

Trello card detailing tests:
https://trello.com/c/2KngFMJL/352-final-tests-before-switching-over-to-use-new-mapit-in-production