Make duplicity config specific to asset-slave-1
Make the configuration for Duplicity's cache directory specific to the
`asset-slave-1` machine.

When I added this config, I accidentally applied it to all machines of
the `asset_slave` type. We only run offsite backups with Duplicity on
`asset-slave-1`, so make the configuration for the Duplicity cache
specific to that machine.