Upgrade saz/rsyslog to 3.2.0
I reviewed this release by release because it was a bit overwhelming otherwise.

## 2.4.0

https://github.com/saz/puppet-rsyslog/compare/v2.3.0...v2.4.0

Nothing of interest to us here.

## 3.0.0

https://github.com/saz/puppet-rsyslog/compare/v2.4.0...v3.0.0

Fixes a bug introduced in 2.4.0 by moving the `preserve_fqdn` param.
We use this param in `govuk::node::s_base` so I've moved it there.
This is the only breaking change in that release.

## 3.1.0

https://github.com/saz/puppet-rsyslog/compare/v3.0.0...v3.1.0

Adds a restart command to the rsyslog service. I think this is ok.

## 3.2.0

https://github.com/saz/puppet-rsyslog/compare/v3.1.0...v3.2.0

Sets `InputFilePersistStateInterval` to 0 which is its default value.

Fixes scope of variables. This is the big one because it helps get us
to having a working future parser.