Restrict inline sed script scope
The inline sed is only necessary for tariff_temporal and slows down all
other databases. I’ve made a mapping file optional per database - if a
.sed file exists it will be used.

Additionally, performing the db name replacement for the entire file is
very expensive. If we anchor the sed script to only run for the first
10 lines inclusive after the CREATE DATABASE line, we can capture all
occurrences of the database name and avoid needlessly running the
script for every line.