Add per-queue stats for rabbitmq.
We achieve this by replacing the existing python script with
a plugin packaged by NYTimes which does per-queue stats
out of the box.

Note the lack of testing - I've not even proved this works
manually so it probably should not be merged for now.

I'm not sure how to examine the collectd output in a vm,
nor whether I am doing the right thing with puppet.

Advice welcome!

@benhyland and @tijmenb