Prevent accidental activation of rummager worker.
Recently the capistrano deployment job ran a task which erroneously enabled
the rummager indexer worker process. This should not be possible, and
could possibly have led to production issues (we got away with it this time).

Tie the presence of the upstart conf file for the worker to the feature flag -
when the flag is off the file is missing so it is not possible to accidentally
start the worker.

 Also tie the appropriate rabbitmq queue config to the same feature.

 @matmoore and @benhyland