Redo "Add per-queue stats for rabbitmq"
This reverts commit b3b46fa595619293dd0ca314b309e493edf986be.

- "Revert the revert of "Add per-queue stats for rabbitmq""
- Add missing depenency that was breaking the dev VM
- Change resources to virtual resources

The previous attempt at this caused some problems in development
because we were running collectd plugins on nodes where we didn't
install the collectd package. On dev, we disable a lot of monitoring
functionality. We now use an explicit `require` dependency and use
virtual resources so that the plugins are only set up when collected
into the collectd::plugins class.