Configure queues and queue bindings for rabbitmq from puppet
We are going to stop handling queue creation and binding in the consumer
processes, since it is too easy to mess up and leave a unintended binding
around.

Instead, we will configure everything in puppet and the consumer process
will error out if the queue it expects is missing.

This is the first step.
Next, we will modify the consumer gem and provide a way to remove
unwanted bindings via puppet (perhaps automatically).

We deliberately omit stagecraft and backdrop_write config, since they do
not use the consumer gem and it is likely to change as performance
platform exits alpha.

See https://trello.com/c/eP2CdnrC/590-fix-rabbitmq-queue-binding

@matmoore and @benhyland